requirements.txt
streamlit
psycopg2-binary
python-dotenv
pandas
json
plotly.express
plotly.graph_objects 